Start working on Sendable / better message builders
