+++ requirements.txt
@@ -1,5 +1,5 @@
 gevent==1.2.2
 holster==1.0.15
-requests==2.13.0
+requests==2.18.1
 six==1.10.0
-websocket-client==0.40.0
+websocket-client==0.44.0
