+++ setup.py
@@ -13,7 +13,7 @@ extras_require = {
     'voice': ['pynacl==1.1.2'],
     'http': ['flask==0.12.2'],
     'yaml': ['pyyaml==3.12'],
-    'music': ['youtube_dl==2017.4.26'],
+    'music': ['youtube_dl==2017.7.2'],
     'performance': ['erlpack==0.3.2', 'ujson==1.35'],
     'sharding': ['gipc==0.6.0'],
     'docs': ['biblio==0.0.2'],
